refactor: move measurement checkbox to a separate file
